[TASK] Add RequireJs functionality
This JavaScript removes any registered action for onblur and onfocus of
the form fields. Furthermore it adds an own function to stop the event
from bubbling up the DOM tree.

It adds the approbate HTML to show the button and registers and onclick
function to switch the form field type.